+++ main.tf
@@ -62,7 +62,7 @@ module "cluster" {
   admin_group_object_ids               = var.admin_group_object_ids
   bootstrap_name                       = local.bootstrap_name
   bootstrap_vm_size                    = local.bootstrap_vm_size
-  logging                              = var.logging
+  logging                              = local.logging
   storage                              = var.storage
   maintenance                          = var.maintenance
   oms_agent                            = var.experimental.oms_agent
@@ -144,7 +144,7 @@ module "core_config" {
 
   dns_resource_group_lookup = var.dns_resource_group_lookup
 
-  logging = merge(var.logging, { control_plane = { log_analytics = merge(var.logging.control_plane.log_analytics, { workspace_id = module.cluster.control_plane_log_analytics_workspace_id }), storage_account = var.logging.control_plane.storage_account } })
+  logging = local.logging
   storage = var.storage
 
   core_services_config = local.core_services_config