+++ modules/cluster/logging.tf
@@ -4,7 +4,7 @@ resource "azurerm_monitor_diagnostic_setting" "workspace" {
   name               = "control-plane-log-analytics"
   target_resource_id = azurerm_kubernetes_cluster.default.id
 
-  log_analytics_workspace_id = var.logging.control_plane.log_analytics.workspace_id
+  log_analytics_workspace_id = coalesce(var.logging.control_plane.log_analytics.workspace_id, var.logging.log_analytics_workspace_config.id)
 
   log_analytics_destination_type = "AzureDiagnostics"
 
@@ -35,7 +35,7 @@ resource "azurerm_monitor_diagnostic_setting" "storage_account" {
   name               = "control-plane-storage-account"
   target_resource_id = azurerm_kubernetes_cluster.default.id
 
-  storage_account_id = var.logging.storage_account_config.id == null ? var.logging.control_plane.storage_account.id : var.logging.storage_account_config.id
+  storage_account_id = coalesce(var.logging.control_plane.storage_account.id, var.logging.storage_account_config.id)
 
   dynamic "enabled_log" {
     for_each = local.storage_account_log_category_types