+++ modules/core-config/modules/fluentd/local.tf
@@ -176,26 +176,26 @@ locals {
     EOT
   }
 
-  azure_route_config = var.azure_storage_output.enabled ? {
-    match  = "**"
-    label  = "@AZURE"
+  azure_storage_nodes_route_config = var.azure_storage_nodes_output.enabled ? {
+    match  = "host.**"
+    label  = "@AZURE_NODES"
     copy   = true
     config = <<-EOT
       <match **>
         @type azurestorage_gen2
-        azure_storage_account         ${regex("[[:alnum:]]+$", var.azure_storage_output.id)}
-        azure_container               ${var.azure_storage_output.container}
+        azure_storage_account         ${regex("[[:alnum:]]+$", var.azure_storage_nodes_output.id)}
+        azure_container               ${var.azure_storage_nodes_output.container}
         azure_instance_msi            ${module.identity.id}
         azure_client_id               ${module.identity.client_id}
         azure_object_key_format       %%{path}/%%{time_slice}_$${chunk_id}.%%{file_extension}
         time_slice_format             %Y%m%d-%H
-        path                          "${join("/", compact([var.azure_storage_output.path_prefix, "kubernetes/${var.cluster_name}/%Y/%m/%d"]))}"
+        path                          "${join("/", compact([var.azure_storage_nodes_output.path_prefix, "kubernetes/${var.cluster_name}/%Y/%m/%d"]))}"
         auto_create_container         true
         store_as                      gzip
         format                        json
         <buffer time>
           @type file
-          path /fluentd/state/azurestorage
+          path /fluentd/state/azurestorage/nodes
           timekey 900
           timekey_wait 60
           timekey_use_utc true
@@ -209,7 +209,40 @@ locals {
     EOT
   } : {}
 
-  route_config = concat(var.loki_output.enabled ? [local.loki_route_config] : [], var.azure_storage_output.enabled ? [local.azure_route_config] : [], length(var.route_config) == 0 || var.debug ? [{
+  azure_storage_workloads_route_config = var.azure_storage_workloads_output.enabled ? {
+    match  = "kube.**"
+    label  = "@AZURE_WORKLOADS"
+    copy   = true
+    config = <<-EOT
+      <match **>
+        @type azurestorage_gen2
+        azure_storage_account         ${regex("[[:alnum:]]+$", var.azure_storage_workloads_output.id)}
+        azure_container               ${var.azure_storage_workloads_output.container}
+        azure_instance_msi            ${module.identity.id}
+        azure_client_id               ${module.identity.client_id}
+        azure_object_key_format       %%{path}/%%{time_slice}_$${chunk_id}.%%{file_extension}
+        time_slice_format             %Y%m%d-%H
+        path                          "${join("/", compact([var.azure_storage_workloads_output.path_prefix, "kubernetes/${var.cluster_name}/%Y/%m/%d"]))}"
+        auto_create_container         true
+        store_as                      gzip
+        format                        json
+        <buffer time>
+          @type file
+          path /fluentd/state/azurestorage/workloads
+          timekey 900
+          timekey_wait 60
+          timekey_use_utc true
+          chunk_limit_size 128MB
+          total_limit_size 8GB
+          overflow_action drop_oldest_chunk
+          retry_type exponential_backoff
+          retry_forever true
+        </buffer>
+      </match>
+    EOT
+  } : {}
+
+  route_config = concat(var.loki_output.enabled ? [local.loki_route_config] : [], var.azure_storage_nodes_output.enabled ? [local.azure_storage_nodes_route_config] : [], var.azure_storage_workloads_output.enabled ? [local.azure_storage_workloads_route_config] : [], length(var.route_config) == 0 || var.debug ? [{
     match  = "**"
     label  = length(var.route_config) > 0 ? "@DEBUG" : "@DEFAULT"
     copy   = length(var.route_config) > 0
@@ -235,6 +268,8 @@ locals {
 ${trimspace(var.filters)}
 %{endif~}
 EOT
+
+  storage_account_ids = distinct(compact(concat(var.azure_storage_nodes_output.enabled ? [var.azure_storage_nodes_output.id] : [], var.azure_storage_workloads_output.enabled ? [var.azure_storage_workloads_output.id] : [])))
 
   service_account_name = local.name